==> NLP_in_python/2-Exploratory-Data-Analysis.ipynb <==
### Exploratory Data Analysis
#### Introduction

#### After the data cleaning step where we put our data into a few standard formats, the next step is to take a look at the data and see if what we're looking at makes sense. Before applying any fancy algorithms, it's always important to explore the data first.

#### When working with numerical data, some of the exploratory data analysis (EDA) techniques we can use include finding the average of the data set, the distribution of the data, the most common values, etc. The idea is the same when working with text data. We are going to find some more obvious patterns with EDA before identifying the hidden patterns with machines learning (ML) techniques. We are going to look at the following for each comedian:

    1.Most common words - find these and create word clouds
    2.Size of vocabulary - look number of unique words and also how quickly someone speaks
    3.Amount of profanity - most common terms

### Most Common Words
#### Analysis

# Read in the document-term matrix
import pandas as pd

data = pd.read_pickle('dtm.pkl')
data = data.transpose()
data.head()

data.columns

# Find the top 30 words said by each comedian
top_dict = {}
for c in data.columns:
    top = data[c].sort_values(ascending=False).head(30)
    top_dict[c] = list(zip(top.index,top.values))
top_dict

# Print the top 15 words said by each comedian
for comedians,top_words in top_dict.items():
    print(comedians)
    print(', '.join([word for word,count in top_words[0:14]]))
    print('---')

#### NOTE: At this point, we could go on and create word clouds. However, by looking at these top words, you can see that some of them have very little meaning and could be added to a stop words list, so let's do just that.

# Look at the most common top words --> add them to the stop word list
from collections import Counter

# Let's first pull out the top 30 words for each comedian
words = []
for comedian in data.columns:
    top = [word for (word,count) in top_dict[comedian]]
    for t in top:
        words.append(t)
words

# Let's aggregate this list and identify the most common words along with how many routines they occur in
Counter(words).most_common()

# If more than half of the comedians have it as a top word, exclude it from the list
add_stop_words = [word for word,count in Counter(words).most_common() if count>6]
add_stop_words
